support facet ranges now like gbfacetint:gblang,0-1,1-2,2-3,3-4,5-6,6-18

+++ Query.h
@@ -377,11 +377,11 @@ class QueryWord {
 	SafeBuf m_synWordBuf;
 
 
-	long m_facetRangeIntA   [MAX_FACET_RANGES];
-	long m_facetRangeIntB   [MAX_FACET_RANGES];
-	long m_facetRangeFloatA [MAX_FACET_RANGES];
-	long m_facetRangeFloatB [MAX_FACET_RANGES];
-	long m_numFacetRanges;
+	long  m_facetRangeIntA   [MAX_FACET_RANGES];
+	long  m_facetRangeIntB   [MAX_FACET_RANGES];
+	float m_facetRangeFloatA [MAX_FACET_RANGES];
+	float m_facetRangeFloatB [MAX_FACET_RANGES];
+	long  m_numFacetRanges;
 
 
 	// what operand bit # is it for doing boolen queries?
